requirements.txt
streamlit==1.55.0
pyalex==0.21
python-dotenv==1.2.2
psycopg2-binary==2.9.11
